Move Platform trait to a separate module
